+++ src/com/maddyhome/idea/vim/action/change/delete/DeleteJoinVisualLinesAction.java
@@ -56,7 +56,7 @@ final public class DeleteJoinVisualLinesAction extends VimCommandAction {
                                          @NotNull Map<Caret, ? extends VimSelection> caretsAndSelections) {
         if (editor.isOneLineMode()) return false;
 
-        if (OptionsManager.INSTANCE.getSmartjoin().isSet()) {
+        if (OptionsManager.INSTANCE.getIdeajoin().isSet()) {
           VimPlugin.getChange().joinViaIdeaBySelections(editor, context, caretsAndSelections);
           return true;
         }