+++ test/org/jetbrains/plugins/ideavim/action/change/delete/DeleteJoinLinesSpacesActionTest.kt
@@ -20,14 +20,14 @@ package org.jetbrains.plugins.ideavim.action.change.delete
 
 import com.maddyhome.idea.vim.command.CommandState
 import com.maddyhome.idea.vim.helper.StringHelper
-import com.maddyhome.idea.vim.option.SmartJoinOptionsData
+import com.maddyhome.idea.vim.option.IdeaJoinOptionsData
 import org.jetbrains.plugins.ideavim.VimOptionTestCase
 import org.jetbrains.plugins.ideavim.VimOptionTestConfiguration
 import org.jetbrains.plugins.ideavim.VimTestOption
 import org.jetbrains.plugins.ideavim.VimTestOptionType
 
-class DeleteJoinLinesSpacesActionTest : VimOptionTestCase(SmartJoinOptionsData.name) {
-  @VimOptionTestConfiguration(VimTestOption(SmartJoinOptionsData.name, VimTestOptionType.TOGGLE, ["true"]))
+class DeleteJoinLinesSpacesActionTest : VimOptionTestCase(IdeaJoinOptionsData.name) {
+  @VimOptionTestConfiguration(VimTestOption(IdeaJoinOptionsData.name, VimTestOptionType.TOGGLE, ["true"]))
   fun `test join with idea`() {
     doTest(StringHelper.parseKeys("J"),
       """
@@ -49,7 +49,7 @@ class DeleteJoinLinesSpacesActionTest : VimOptionTestCase(SmartJoinOptionsData.n
       CommandState.SubMode.NONE)
   }
 
-  @VimOptionTestConfiguration(VimTestOption(SmartJoinOptionsData.name, VimTestOptionType.TOGGLE, ["true"]))
+  @VimOptionTestConfiguration(VimTestOption(IdeaJoinOptionsData.name, VimTestOptionType.TOGGLE, ["true"]))
   fun `test join with idea with count`() {
     doTest(StringHelper.parseKeys("3J"),
       """
@@ -70,7 +70,7 @@ class DeleteJoinLinesSpacesActionTest : VimOptionTestCase(SmartJoinOptionsData.n
       CommandState.SubMode.NONE)
   }
 
-  @VimOptionTestConfiguration(VimTestOption(SmartJoinOptionsData.name, VimTestOptionType.TOGGLE, ["true"]))
+  @VimOptionTestConfiguration(VimTestOption(IdeaJoinOptionsData.name, VimTestOptionType.TOGGLE, ["true"]))
   fun `test join with idea with large count`() {
     doTest(StringHelper.parseKeys("10J"),
       """