+++ test/org/jetbrains/plugins/ideavim/action/change/delete/DeleteJoinVisualLinesSpacesActionTest.kt
@@ -20,14 +20,14 @@ package org.jetbrains.plugins.ideavim.action.change.delete
 
 import com.maddyhome.idea.vim.command.CommandState
 import com.maddyhome.idea.vim.helper.StringHelper.parseKeys
-import com.maddyhome.idea.vim.option.SmartJoinOptionsData
+import com.maddyhome.idea.vim.option.IdeaJoinOptionsData
 import org.jetbrains.plugins.ideavim.VimOptionTestCase
 import org.jetbrains.plugins.ideavim.VimOptionTestConfiguration
 import org.jetbrains.plugins.ideavim.VimTestOption
 import org.jetbrains.plugins.ideavim.VimTestOptionType
 
-class DeleteJoinVisualLinesSpacesActionTest : VimOptionTestCase(SmartJoinOptionsData.name) {
-  @VimOptionTestConfiguration(VimTestOption(SmartJoinOptionsData.name, VimTestOptionType.TOGGLE, ["true"]))
+class DeleteJoinVisualLinesSpacesActionTest : VimOptionTestCase(IdeaJoinOptionsData.name) {
+  @VimOptionTestConfiguration(VimTestOption(IdeaJoinOptionsData.name, VimTestOptionType.TOGGLE, ["true"]))
   fun `test join via idea`() {
     doTest(parseKeys("VjJ"),
       """